--- backend/database/versions/31_253d1d67ea4a_remove_curator_tag.py
@@ -1,30 +1,0 @@
-"""remove_curator_tag
-
-Revision ID: 253d1d67ea4a
-Revises: 30_26c54abcaac9
-Create Date: 2022-09-07 13:39:39.401203
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "31_253d1d67ea4a"
-down_revision = "30_26c54abcaac9"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint("_dataset__collection_id_curator_tag", "dataset", type_="unique")
-    op.drop_column("dataset", "curator_tag")
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("dataset", sa.Column("curator_tag", sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.create_unique_constraint("_dataset__collection_id_curator_tag", "dataset", ["collection_id", "curator_tag"])
-    # ### end Alembic commands ###
